.savechat moved to utility module
